Remove and refactor utils
