+++ src/datasets/summarizations/utils/commons.ts
@@ -1,4 +1,5 @@
 import { Summary, QuerySummariesFactory } from '../types';
+import { XYPoint, NumPoint } from '../../metas/types';
 
 export function cacheSummaries(f: () => Summary[]): () => Summary[] {
   let cache: Summary[] | null = null;
@@ -24,4 +25,43 @@ export function combineQuerySummariesFactories<PointT>(
     const summariesFlat = ([] as Summary[]).concat(...summaries);
     return summariesFlat;
   };
+}
+
+function pointToPair<X, Y>(p: XYPoint<X, Y>): [X, Y] {
+  return [p.x, p.y];
+}
+
+interface ChartAxisLimit {
+  min?: number;
+  max?: number;
+}
+
+export function normalizePointsX<T>(points: XYPoint<number, T>[], xlim: ChartAxisLimit = {}): XYPoint<number, T>[] {
+  const xValues = points.map(({ x }) => x);
+  const {
+    min: xmin = Math.min(...xValues),
+    max: xmax = Math.max(...xValues),
+  } = xlim;
+
+  return points.map(({ x, y }) => ({
+    x: (x - xmin) / (xmax - xmin),
+    y,
+  }));
+}
+
+export function normalizePointsY<T>(points: XYPoint<T, number>[], ylim: ChartAxisLimit = {}): XYPoint<T, number>[] {
+  const yValues = points.map(({ y }) => y);
+  const {
+    min: ymin = Math.min(...yValues),
+    max: ymax = Math.max(...yValues),
+  } = ylim;
+
+  return points.map(({ x, y }) => ({
+    x,
+    y: (y - ymin) / (ymax - ymin) * 5 / 8,
+  }));
+}
+
+export function normalizePoints(points: NumPoint[], xlim: ChartAxisLimit = {}, ylim: ChartAxisLimit = {}): NumPoint[] {
+  return normalizePointsY(normalizePointsX(points, xlim), ylim);
 }
